+++ components/bottombar/impl/src/main/java/com/flipperdevices/bottombar/impl/api/BottomBarDecomposeComponentImpl.kt
@@ -138,7 +138,8 @@ class BottomBarDecomposeComponentImpl @AssistedInject constructor(
         is BottomBarTabConfig.Tools -> toolsScreenFactory(
             componentContext = componentContext,
             deeplink = config.deeplink,
-            onBack = { navigation.popOr(onBack::invoke) }
+            onBack = { navigation.popOr(onBack::invoke) },
+            onDeeplink = ::handleDeeplink
         )
 
         is BottomBarTabConfig.Apps -> fapHubScreenFactory(