+++ components/remote-controls/grid/remote/impl/src/main/kotlin/com/flipperdevices/remotecontrols/impl/grid/remote/presentation/decompose/internal/RemoteGridComponentImpl.kt
@@ -10,7 +10,7 @@ import com.flipperdevices.keyedit.api.NotSavedFlipperFile
 import com.flipperdevices.keyedit.api.NotSavedFlipperKey
 import com.flipperdevices.remotecontrols.api.DispatchSignalApi
 import com.flipperdevices.remotecontrols.api.SaveTempSignalApi
-import com.flipperdevices.remotecontrols.api.model.ServerRemoteControlParam
+import com.flipperdevices.remotecontrols.grid.remote.api.model.ServerRemoteControlParam
 import com.flipperdevices.remotecontrols.impl.grid.remote.presentation.decompose.RemoteGridComponent
 import com.flipperdevices.remotecontrols.impl.grid.remote.presentation.mapping.GridComponentStateMapper
 import com.flipperdevices.remotecontrols.impl.grid.remote.presentation.viewmodel.ConnectionViewModel
@@ -101,6 +101,8 @@ class RemoteGridComponentImpl @AssistedInject constructor(
         dispatchSignalApi.dismissBusyDialog()
     }
 
+    private fun FlipperFilePath.toNonTempPath() = copy(folder = folder.replace("/temp", ""))
+
     override fun save() {
         val rawRemotes = remoteGridViewModel.getRawRemotesContent() ?: return
         val rawUi = remoteGridViewModel.getRawPagesContent() ?: return
@@ -110,7 +112,7 @@ class RemoteGridComponentImpl @AssistedInject constructor(
                 FlipperFilePath(
                     folder = param.extTempFolderPath,
                     nameWithExtension = param.nameWithExtension
-                ),
+                ).toNonTempPath(),
                 content = FlipperKeyContent.RawData(rawRemotes.toByteArray())
             ),
             additionalFiles = listOf(
@@ -118,7 +120,7 @@ class RemoteGridComponentImpl @AssistedInject constructor(
                     path = FlipperFilePath(
                         folder = param.extTempFolderPath,
                         nameWithExtension = param.uiFileNameWithExtension
-                    ),
+                    ).toNonTempPath(),
                     content = FlipperKeyContent.RawData(rawUi.toByteArray())
                 )
             ),