+++ components/remote-controls/main/api/src/main/kotlin/com/flipperdevices/remotecontrols/api/RemoteControlsScreenDecomposeComponent.kt
@@ -1,6 +1,7 @@
 package com.flipperdevices.remotecontrols.api
 
 import com.arkivanov.decompose.ComponentContext
+import com.flipperdevices.deeplink.model.Deeplink
 import com.flipperdevices.ui.decompose.CompositeDecomposeComponent
 import com.flipperdevices.ui.decompose.DecomposeOnBackParameter
 
@@ -9,6 +10,7 @@ abstract class RemoteControlsScreenDecomposeComponent<C : Any> : CompositeDecomp
         operator fun invoke(
             componentContext: ComponentContext,
             onBack: DecomposeOnBackParameter,
+            onDeeplink: (Deeplink.BottomBar) -> Unit
         ): RemoteControlsScreenDecomposeComponent<*>
     }
 }