+++ components/rootscreen/impl/src/main/kotlin/com/flipperdevices/rootscreen/impl/api/RootDecomposeComponentImpl.kt
@@ -22,8 +22,6 @@ import com.flipperdevices.faphub.screenshotspreview.api.ScreenshotsPreviewDecomp
 import com.flipperdevices.firstpair.api.FirstPairApi
 import com.flipperdevices.firstpair.api.FirstPairDecomposeComponent
 import com.flipperdevices.keyscreen.api.KeyScreenDecomposeComponent
-import com.flipperdevices.remotecontrols.api.ConfigureGridDecomposeComponent
-import com.flipperdevices.remotecontrols.api.model.ServerRemoteControlParam
 import com.flipperdevices.rootscreen.api.RootDecomposeComponent
 import com.flipperdevices.rootscreen.impl.deeplink.RootDeeplinkHandler
 import com.flipperdevices.rootscreen.model.RootScreenConfig
@@ -58,7 +56,6 @@ class RootDecomposeComponentImpl @AssistedInject constructor(
     private val keyScreenFactory: KeyScreenDecomposeComponent.Factory,
     private val screenshotsPreviewFactory: ScreenshotsPreviewDecomposeComponent.Factory,
     private val changelogScreenDecomposeFactory: ChangelogScreenDecomposeComponent.Factory,
-    private val serverRemoteControlFactory: ConfigureGridDecomposeComponent.Factory
 ) : RootDecomposeComponent, ComponentContext by componentContext {
     private val scope = coroutineScope(FlipperDispatchers.workStealingDispatcher)
     private val navigation = StackNavigation<RootScreenConfig>()
@@ -130,15 +127,6 @@ class RootDecomposeComponentImpl @AssistedInject constructor(
             componentContext = componentContext,
             updateRequest = config.updateRequest,
             onBack = this::internalOnBack
-        )
-
-        is RootScreenConfig.ServerRemoteControl -> serverRemoteControlFactory(
-            componentContext = componentContext,
-            param = ServerRemoteControlParam(
-                infraredFileId = config.infraredFileId,
-                remoteName = config.remoteName
-            ),
-            onBack = this::internalOnBack,
         )
     }