+++ CHANGELOG.md
@@ -60,6 +60,7 @@ Attention: don't forget to add the flag for F-Droid before release
 - [FIX] Replace /any/ to /ext/, because /any/ is deprecated
 - [FIX] Remove share from remote controls
 - [FIX] Remove "RPC" argument from screen streaming request
+- [FIX] Fix remote-controls duplication ir files
 - [CI] Fix merge-queue files diff
 - [CI] Add https://github.com/LionZXY/detekt-decompose-rule
 - [CI] Enabling detekt module for android and kmp modules